model/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
torchaudio
Pillow
timm
